Controllers, Vistas, autenticacion, layout

+++ CarritoCompras-NT1/Controllers/CarritoItemsController.cs
@@ -1,15 +1,16 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using CarritoCompras_NT1.DataBase;
+using CarritoCompras_NT1.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
-using CarritoCompras_NT1.DataBase;
-using CarritoCompras_NT1.Models;
+using System;
+using System.Linq;
+using System.Security.Claims;
+using System.Threading.Tasks;
 
 namespace CarritoCompras_NT1.Controllers
 {
+    //[Authorize(Roles = nameof(Rol.Cliente))]
     public class CarritoItemsController : Controller
     {
         private readonly Contexto _context;
@@ -20,10 +21,39 @@ namespace CarritoCompras_NT1.Controllers
         }
 
         // GET: CarritoItems
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(Guid? id)
         {
-            var contexto = _context.CarritoItems.Include(c => c.Carrito).Include(c => c.Producto);
-            return View(await contexto.ToListAsync());
+            
+            var clienteId = Guid.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+            Carrito carrito = null;
+            if (id == null)
+            {
+                carrito = _context.Carritos.FirstOrDefault(car => car.ClienteID == clienteId && car.Activo);
+            }
+            else
+            {
+                carrito = _context.Carritos.FirstOrDefault(car => car.ClienteID == clienteId && car.Id == id);
+            }
+
+            if (carrito== null)
+            {
+                return NotFound();
+            }
+            
+
+            var items = _context.CarritoItems.Include(c => c.Carrito)
+                .Include(c => c.Producto)
+                .Where(c => c.CarritoID == carrito.Id);
+
+            float total = 0;
+            foreach (CarritoItem item in items)
+            {
+                total += item.Subtotal;
+            }
+
+            ViewBag.CarritoId = carrito.Id;
+            ViewBag.Total = total;
+            return View(await items.ToListAsync());
         }
 
         // GET: CarritoItems/Details/5
@@ -45,6 +75,8 @@ namespace CarritoCompras_NT1.Controllers
 
             return View(carritoItem);
         }
+
+        /*
 
         // GET: CarritoItems/Create
         public IActionResult Create()
@@ -128,6 +160,8 @@ namespace CarritoCompras_NT1.Controllers
             return View(carritoItem);
         }
 
+        */
+
         // GET: CarritoItems/Delete/5
         public async Task<IActionResult> Delete(Guid? id)
         {
@@ -153,11 +187,29 @@ namespace CarritoCompras_NT1.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(Guid id)
         {
-            var carritoItem = await _context.CarritoItems.FindAsync(id);
+            CarritoItem carritoItem = _context.CarritoItems.FirstOrDefault(item => item.Id == id);
+            if (carritoItem == null)
+            {
+                return NotFound();
+            }
+            Carrito carrito = _context.Carritos.FirstOrDefault(c => c.Id == carritoItem.CarritoID);
+            if (carrito == null)
+            {
+                return NotFound();
+            }
+
+            // Descontamos el subtotal del item eliminado del subtotal del carrito
+            carrito.Subtotal -= carritoItem.Subtotal;
+
+            _context.Update(carrito);
             _context.CarritoItems.Remove(carritoItem);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
+
+
+
+
 
         private bool CarritoItemExists(Guid id)
         {
